requirements.txt
fastapi~=0.115.5
pydantic~=2.9.2
pandas~=2.2.3
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
pytest~=8.3.3
httpx~=0.27.2